# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for the Assignment001 API.",
        "title": "Assignment001 API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/echo": {
            "get": {
                "description": "Sending a get request to the echo endpoint should always return a response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Global"
                ],
                "summary": "Uptime endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/assignment001.echoResponse"
                        }
                    }
                }
            }
        },
        "/v1.0/tweet/get/stream": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Handles getting filtered stream tweets of a special user with special rules",
                "tags": [
                    "V1 Twitter"
                ],
                "summary": "get filter stream tweets",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/v1.0/tweet/get/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Handles getting a special tweet",
                "tags": [
                    "V1 Twitter"
                ],
                "summary": "get tweet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "a valid tweet id value",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/twitterv1.GetTweetResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "assignment001.echoResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "twitterv1.Data": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "twitterv1.GetTweetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/twitterv1.Data"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}